pop/pyproject.toml
[tool.poetry]
name = "page-object-playwright"
version = "0.1.0"
description = ""
authors = ["Rodionchik <insert_into_u_ass@cum.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pytest-playwright = "^0.4.3"
pytest-asyncio = "^0.21.1"
allure-pytest = "^2.13.2"
playwright = "^1.40.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
